Move current_html to WpHTMLUtil
